+++ mapper/test/test.constructors.dart
@@ -164,28 +164,6 @@ class IgnoredFieldClassWoConstructor {
 }
 
 @jsonSerializable
-class CropArea {
-  num? top;
-  num? left;
-  num? right;
-  num? bottom;
-}
-
-@jsonSerializable
-class ImmutableDefault {
-  @JsonProperty(
-      defaultValue: {'top': 0.0, 'left': 1.0, 'right': 1.0, 'bottom': 0.0})
-  final CropArea? cropArea;
-
-  @JsonProperty(defaultValue: 1)
-  final int? id;
-  final String? name;
-  final Car? car;
-
-  const ImmutableDefault({this.cropArea, this.id, this.name, this.car});
-}
-
-@jsonSerializable
 @Json(processAnnotatedMembersOnly: true)
 class ImmutableDefault2 {
   @JsonProperty(defaultValue: 1)
@@ -424,17 +402,6 @@ void testConstructors() {
       expect(target, '{"firstName":"Bob"}');
     });
 
-    test('processAnnotatedMembersOnly global option', () {
-      // given
-      final instance = ImmutableDefault();
-      // when
-      final target = JsonMapper.serialize(instance,
-          SerializationOptions(indent: '', processAnnotatedMembersOnly: true));
-      // then
-      expect(target,
-          '{"cropArea":{"top":0.0,"left":1.0,"right":1.0,"bottom":0.0},"id":1}');
-    });
-
     test('processAnnotatedMembersOnly class annotation option', () {
       // given
       final instance = ImmutableDefault2();
@@ -489,46 +456,6 @@ void testConstructors() {
       final ic = JsonMapper.deserialize<Immutable>(immutableJson);
       // then
       expect(JsonMapper.serialize(ic), immutableJson);
-    });
-
-    test('Serialize Immutable class with DefaultValue provided', () {
-      // given
-      final immutableJson = '''{
- "cropArea": {
-  "top": 0.0,
-  "left": 1.0,
-  "right": 1.0,
-  "bottom": 0.0
- },
- "id": 1,
- "name": "Bob",
- "car": {
-  "modelName": "Audi",
-  "color": "green"
- }
-}''';
-
-      final json = '''{
- "name": "Bob",
- "car": {
-  "modelName": "Audi",
-  "color": "Green"
- }
-}''';
-      final i = ImmutableDefault(name: 'Bob', car: Car('Audi', Color.green));
-
-      // when
-      final targetJson = JsonMapper.serialize(i);
-      final target = JsonMapper.deserialize<ImmutableDefault>(json)!;
-
-      // then
-      expect(targetJson, immutableJson);
-
-      expect(target.id, 1);
-      expect(target.cropArea, TypeMatcher<CropArea>());
-      expect(target.cropArea!.left, 1);
-      expect(target.cropArea!.right, 1);
-      expect(target.cropArea!.bottom, 0);
     });
 
     test('Back referencing to parent/owner object', () {